requirements.txt
# WebWeaver 依赖包

# 核心依赖
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
beautifulsoup4==4.12.2

# 数据处理
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# 文本处理
nltk==3.8.1
spacy==3.7.2
textstat==0.7.3

# 配置管理
pyyaml==6.0.1
python-dotenv==1.0.0

# 日志和监控
loguru==0.7.2

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# 开发工具
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# 可选依赖（用于特定功能）
# openai==1.3.0  # 用于OpenAI API
# anthropic==0.7.8  # 用于Anthropic API
# langchain==0.0.350  # 用于LangChain集成
# chromadb==0.4.18  # 用于向量数据库
# redis==5.0.1  # 用于缓存
